+++ CHANGELOG.md
@@ -4,3 +4,10 @@ This project uses [Semantic Versioning](http://semver.org/).
 ## 0.0.1 - 2015-12-10
 
 Initial version.
+
+
+## 0.0.2 - 2015-12-10
+
++ report workers by queue name even if the queues themselves do not exist
++ improve logging
++ add --debug and --no-cloudwatch options